refactor(HomeScreen): 替换脚本运行逻辑并优化配置实体
• 在HomeScreen.kt中注释原有脚本运行逻辑，替换为RunScript.testOcr()
• 在ConfigData.kt中移除冗余字段，新增Bitmap和包名字段
• 在DetectResult.kt中将label类型从Int改为Short

本次重构旨在简化配置实体结构并优化脚本运行逻辑，提升代码可维护性。

+++ app/src/main/java/com/smart/autodaily/data/entity/DetectResult.kt
@@ -1,6 +1,6 @@
 package com.smart.autodaily.data.entity
 data class DetectResult(
-    val label: Int,
+    val label: Short,
     val prob: Float,
     val rect: Rect,
     val xCenter: Float,
